fix: ensure react types are only imported if a react renderer is used
