+++ src/server/pure.ts
@@ -2,10 +2,11 @@ import ReactDOMServer from 'react-dom/server'
 import ReactDOM from 'react-dom'
 import { act } from 'react-dom/test-utils'
 
-import { RendererProps, ReactRendererOptions } from '../types'
+import { RendererProps } from 'types'
+import { ReactRendererOptions } from '../react/types'
 
 import { createRenderHook, cleanup, addCleanup, removeCleanup } from '../core'
-import toRender from '../helpers/toRender'
+import { createTestHarness } from '../react/createTestHarness'
 
 function createServerRenderer<TProps, TResult>(
   testHookProps: RendererProps<TProps, TResult>,
@@ -13,7 +14,7 @@ function createServerRenderer<TProps, TResult>(
 ) {
   const container = document.createElement('div')
 
-  const testHook = toRender(testHookProps, wrapper, false)
+  const testHook = createTestHarness(testHookProps, wrapper, false)
 
   let renderProps: TProps | undefined
   let hydrated = false
